few fixes in debug mode

+++ RunCPM/cpu.h
@@ -30,13 +30,13 @@ extern int32_t _cpu_step;
 
 #define CPU_LOW_DIGIT(x)            ((x) & 0xf)
 #define CPU_HIGH_DIGIT(x)           (((x) >> 4) & 0xf)
-#define CPU_REG_GET_LOW(x)         ((x) & 0xff)
-#define CPU_REG_GET_HIGH(x)        (((x) >> 8) & 0xff)
+#define CPU_REG_GET_LOW(x)         (uint16_t)((x) & 0xff)
+#define CPU_REG_GET_HIGH(x)        (uint16_t)(((x) >> 8) & 0xff)
 
 #define CPU_REG_SET_LOW(x, v)  x = (((x) & 0xff00) | ((v) & 0xff))
 #define CPU_REG_SET_HIGH(x, v) x = (((x) & 0xff) | (((v) & 0xff) << 8))
 
-#define CPU_WORD16(x)	((x) & 0xffff)
+#define CPU_WORD16(x)	((uint16_t)((x) & 0xffff))
 
 #ifdef __cplusplus
 extern "C"
